Coming middlewares into 1 onboarding middleware
